requirements.txt
six
PyYAML
jinja2
